requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1
alt-profanity-check==1.3.1
playwright==1.40.0
openai==1.6.1
backoff==2.2.1
